% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{b_samp}
\alias{b_samp}
\title{Sample subject specific random effects}
\usage{
b_samp(Omega, beta, R, theta, X, q, Z, ZtZ, N)
}
\arguments{
\item{Omega}{current covariance matrix}

\item{beta}{current fixed effect coefficients vector}

\item{X}{design matrix model parameters (differs per person)}

\item{q}{dimension Z(number of random effects)}

\item{Z}{design matrix for random effects (differs per person)}

\item{ZtZ}{transpose(Z)*Z}

\item{N}{sample size at second level}

\item{Y}{current outcome vector (Y.I=cos(theta)*R, Y.II=sin(theta)*R)}
}
\description{
Sample subject specific random effects
}
\keyword{internal}
