% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{betaBlock.fRI}
\alias{betaBlock.fRI}
\title{Sample fixed effect coefficients in a Random Intercept model}
\usage{
betaBlock.fRI(Omega, Y, X, Z, p, A, N)
}
\arguments{
\item{Omega}{current covariance matrix}

\item{Y}{current outcome vector (Y.I=cos(theta)*R, Y.II=sin(theta)*R)}

\item{X}{design matrix model parameters (differs per person)}

\item{Z}{design matrix for random effects (differs per person)}

\item{p}{dimension X (number of columns/variables+indicator variables)}

\item{A}{prior variance of fixed effect coefficients}

\item{N}{sample size at second level}
}
\description{
Sample fixed effect coefficients in a Random Intercept model
}
