% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classfunctions.R
\name{BFc.bpnme}
\alias{BFc.bpnme}
\title{Bayes Factors for a Bayesian circular mixed-effects model}
\usage{
\method{BFc}{bpnme}(object, hypothesis, type = "anchor")
}
\arguments{
\item{object}{a \code{bpnme object} obtained from the function
\code{bpnme()}.}

\item{hypothesis}{the inequality constrained hypothesis to test.}

\item{type}{type of hypothesis to test c("anchor", "isotropic").}
}
\value{
Bayes Factors for inequality constrained hypotheses on mean differences.
}
\description{
Outputs Bayes Factors for inequality constrained hypotheses on the circular
differences between several levels of a categorical variable and the baseline.
}
