% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classfunctions.R
\name{BFc}
\alias{BFc}
\title{Bayes Factors}
\usage{
BFc(object, hypothesis, type = "anchor")
}
\arguments{
\item{object}{an object used to select a method}

\item{hypothesis}{the inequality constrained hypothesis to test}

\item{type}{type of hypothesis to test c("anchor", "isotropic")}

\item{...}{further arguments passed to or from other methods}
}
\description{
\code{BF} gives bayes factors for inequality constrained hypotheses on
circular mean differences
}
\details{
the methods \link[pnreg]{BFc.bpnr} and
  \link[pnreg]{BFc.bpnme} have their own help page
}
