% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{bpnme}
\alias{bpnme}
\title{Fit a Baysian circular mixed-effects model}
\source{
Cremers, J., Mainhard, M.T. & Klugkist, I. (2017). Assessing a
  Bayesian Embedding Approach to Circular Regression Models. Manuscript under
  review.

Cremers, J., & Klugkist, I. (2017). How to analyze circular data: A
  tutorial for projected normal regression models. Working paper.

Cremers, J., Mainhard, M.T. & Klugkist, I. (2017). Longitudinal
  circular modelling of circumplex measurements for teacher behavior. Working
  paper.

Presnell, B., Morrison, S.P. & Littell, R.C. (1998). Projected
  multivariate linear models for directional data. Journal of the Americal
  Statistical Association, 93 (443), 1068 - 1077.
}
\usage{
bpnme(pred.I, data, pred.II = pred.I, its = 1000, burn = 1, n.lag = 1,
  seed = 101)
}
\arguments{
\item{pred.I}{model equation for effects of component 1.}

\item{data}{the dataframe used for analysis.}

\item{pred.II}{model equation for effects of component 2.}

\item{its}{output iterations of the MCMC sampler.}

\item{burn}{number of burn-in iterations.}

\item{n.lag}{amount of lag for the iterations and burn-in.}

\item{seed}{user-specified random seed.}
}
\description{
This function fits a Bayesian circular mixed-effects model based on the
projected normal distribution.
}
\details{
Because the model is based on the projected normal distribution, a
  model equation has to be given for the fixed and random effects of thetwo
  components. By default the model equation of the second component
  \code{pred.II} is set to be equal to that of the
  first component. For more information about the projected normal
  distribution see Presnell, Morrisson & Littell (1998).

  A tutorial on how to use this function can be found in Cremers & Klugkist
  (2017, working paper). More details on the sampling algorithm and
  interpretation of the coefficients from the model can be found in Cremers,
  Mainhard & Klugkist (2017, working paper).
}
\examples{
library(bpnreg)
bpnme(Error.rad ~ Maze + Trial.type + (1|Subject), Maps, its = 100)

}
