% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cat_check}
\alias{cat_check}
\title{Check whether a variable is categorical}
\usage{
cat_check(x)
}
\arguments{
\item{x}{vector containing values of a variable.}
}
\value{
TRUE, if the vector only contains 0 and 1, FALSE is the vector
  contains other values.
}
\description{
\code{cat_check} checks whether a vector contains only 0 and 1 and thus is a
dummy variable.
}
