% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{circ_coef}
\alias{circ_coef}
\title{Compute circular coefficients from linear coefficients}
\usage{
circ_coef(a1, a2, b1, b2)
}
\arguments{
\item{a1}{intercept of the first linear component.}

\item{a2}{intercept of the second linear component.}

\item{b1}{slope of the first linear component.}

\item{b2}{slope of the second linear component.}
}
\value{
A dataframe containing the coordinates of the inflection point of a
  circular effect, the slope at the inflection point and the unsigned and
  signed shortest distance to the origin.
}
\description{
\code{circ_coef} computes the coordinates of the inflection point of
a circular effect, the slope at the inflection point and the unsigned and
signed shortest distance to the origin.
}
