% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{circ_coef_rcpp}
\alias{circ_coef_rcpp}
\title{Compute circular coefficients}
\usage{
circ_coef_rcpp(a1, a2, b1, b2)
}
\arguments{
\item{a1}{intercept estimate of component I.}

\item{a2}{intercept estimate of component I.}

\item{b1}{slope estimate of component I.}

\item{b2}{slope estimate of component I.}
}
\description{
Compute circular coefficients
}
