% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classfunctions.R
\name{coef_lin}
\alias{coef_lin}
\title{Linear coefficients}
\usage{
coef_lin(object)
}
\arguments{
\item{object}{an object used to select a method}
}
\description{
\code{coef_lin} gives posterior summaries of the linear coefficients
}
\details{
the methods \link[bpnreg]{coef_lin.bpnr} and
  \link[bpnreg]{coef_lin.bpnme} have their own help page
}
\examples{
library(bpnreg)
fit.Motor <- bpnr(pred.I = Phaserad ~ 1 + Cond, data = Motor,
its = 100, burn = 10, n.lag = 3)
coef_lin(fit.Motor)

}
