% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classfunctions.R
\name{coef_ran}
\alias{coef_ran}
\title{Random effect variances}
\usage{
coef_ran(object, type = "linear")
}
\arguments{
\item{object}{an object used to select a method}

\item{type}{one of c("linear", "circular") to get either the
linear or circular variances of the random effects}
}
\description{
\code{coef_ran} gives posterior summaries of the circular or linear random effect variances
}
\details{
the method \link[pnreg]{coef_ran.bpnme} has its own help page
}
