% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{Dbd}
\alias{Dbd}
\title{Compute utmu}
\usage{
Dbd(t, mu1, mu2)
}
\arguments{
\item{t}{current outcome value}

\item{mu1}{current predicted linear mean of the first component}

\item{mu2}{current predicted linear mean of the second component}
}
\description{
Compute utmu
}
\keyword{internal}
