% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DIC_reg}
\alias{DIC_reg}
\title{Compute Model Fit Measures Regression Model}
\usage{
DIC_reg(Output, X1, X2)
}
\arguments{
\item{Output}{output from the circular regression function Regression()}

\item{X1}{model matrix for the first component}

\item{X2}{model matrix for the second component}
}
\description{
Compute Model Fit Measures Regression Model
}
\keyword{internal}
