% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classfunctions.R
\name{fit}
\alias{fit}
\title{Model fit}
\usage{
fit(object)
}
\arguments{
\item{object}{an object used to select a method.}
}
\description{
\code{fit} gives several model fit statistics.
}
\details{
the methods \link[bpnreg]{fit.bpnr} and
  \link[bpnreg]{fit.bpnme} have their own help page.
}
\examples{
library(bpnreg)
fit.Motor <- bpnr(pred.I = Phaserad ~ 1 + Cond, data = Motor,
its = 100, burn = 10, n.lag = 3)
fit(fit.Motor)


}
