% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hmodeC}
\alias{hmodeC}
\title{Estimate the mode by finding the highest posterior density interval}
\usage{
hmodeC(x, cip)
}
\arguments{
\item{x}{a  sample from which to estimate the interval}

\item{cip}{bandwidth for the algorithm, ranging from 0 to 1}
}
\value{
a scalar containing the estimate of the mode
}
\description{
Estimate the mode by finding the highest posterior density interval
}
