% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hmodeciC}
\alias{hmodeciC}
\title{Find the highest density interval of a circular variable}
\usage{
hmodeciC(x, cip)
}
\arguments{
\item{x}{a  sample from which to estimate the interval}

\item{cip}{bandwidth for the algorithm, ranging from 0 to 1}
}
\value{
a vector of length 2 containing the lower and upper bound of the interval
}
\description{
Find the highest density interval of a circular variable
}
