% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hpd_est}
\alias{hpd_est}
\title{Compute the 95 percent HPD of a vector of linear data}
\usage{
hpd_est(x)
}
\arguments{
\item{x}{a vector of linear data}
}
\description{
Compute the 95 percent HPD of a vector of linear data
}
\examples{
library(bpnreg)
hpd_est(Motor$AvAmp)

}
