% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hpd_est_circ}
\alias{hpd_est_circ}
\title{Compute the 95 percent HPD of a vector of circular data}
\usage{
hpd_est_circ(x)
}
\arguments{
\item{x}{a vector of circular data}
}
\description{
Compute the 95 percent HPD of a vector of circular data
}
\examples{
library(bpnreg)
hpd_est_circ(subset(Motor, Cond = "exp")$Phaserad)

}
