% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4code.R
\name{isBar}
\alias{isBar}
\title{isBars function from lme4 package}
\source{
Douglas Bates, Martin Maechler, Ben Bolker, Steve Walker (2015).
 Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical
 Software, 67(1),
 1-48.\href{https://www.jstatsoft.org/article/view/v067i01/0}{doi:10.18637/jss.v067.i01}.
}
\usage{
isBar(term)
}
\arguments{
\item{term}{see lme4 package documentation}
}
\description{
Function to help divide input formula of bpnme object into a fixed and random part
}
