% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lik}
\alias{lik}
\title{Compute the Likelihood of the PN distribution}
\usage{
lik(X1, X2, theta, b1, b2, n)
}
\arguments{
\item{X1}{the model matrix of the first component}

\item{X2}{the model matrix of the second component}

\item{theta}{a circular outcome value}

\item{b1}{estimated linear coefficients of the first component}

\item{b2}{estimated linear coefficients of the second component}

\item{n}{sample size}
}
\description{
Compute the Likelihood of the PN distribution
}
