% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lik_me}
\alias{lik_me}
\title{Compute the Likelihood of the PN distribution (mixed effects)}
\usage{
lik_me(
  theta_cos,
  theta_sin,
  X1,
  X2,
  Z1,
  Z2,
  beta1,
  beta2,
  b1,
  b2,
  N,
  pred,
  iteration
)
}
\arguments{
\item{theta_cos}{A List with the cosine of the circular dependent variable.}

\item{theta_sin}{A List with the sine of the circular dependent variable.}

\item{X1}{A list of fixed effect model matrices for component I.}

\item{X2}{A list of fixed effect model matrices for component II.}

\item{Z1}{A list of random effect model matrices for component I.}

\item{Z2}{A list of random effect model matrices for component II.}

\item{beta1}{estimated fixed effect coefficients of the first component}

\item{beta2}{estimated fixed effect coefficients of the second component}

\item{b1}{estimated random effect coefficients of the first component}

\item{b2}{estimated random effect coefficients of the second component}

\item{N}{sample size at second level}

\item{pred}{An empty list for likelihood computation.}

\item{iteration}{iteration number}
}
\description{
Compute the Likelihood of the PN distribution (mixed effects)
}
