% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lik_reg}
\alias{lik_reg}
\title{Compute the Likelihood of the PN distribution (regression)}
\usage{
lik_reg(X1, X2, theta, beta1, beta2, n)
}
\arguments{
\item{X1}{the model matrix of the first component}

\item{X2}{the model matrix of the second component}

\item{theta}{a circular outcome value}

\item{beta1}{estimated linear coefficients of the first component}

\item{beta2}{estimated linear coefficients of the second component}

\item{n}{sample size}
}
\description{
Compute the Likelihood of the PN distribution (regression)
}
