% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Maps}
\alias{Maps}
\title{The geometry of humans' knowledge of navigation space.}
\format{A data frame with 160 rows and 7 variables:
\describe{
  \item{Subject}{a numeric variable indicating the participant number}
  \item{Trial.no}{a numeric variable indicating the trial number of the
  target a participant had to locate (1-8)}
  \item{Maze}{a between subjects factor variable indicating the type of maze
  a participant was in (0 = Euclidean, 1 = non-Euclidean)}
  \item{Trial.type}{a within subjects factor indicating the type of target (0
  = standard, 1 = probe)}
   \item{Error}{a numeric variable containing the angular error in degrees}
  \item{Learn}{a numeric variable indicating the number of trials a
  participant completed in the training phase}
  \item{L.c}{mean centered Learning}
  \item{Error.rad}{a numeric variable containing the angular error in radians}
}}
\source{
\url{https://doi.org/10.1016/j.cognition.2017.05.020}
}
\usage{
Maps
}
\description{
A dataset from a research by Warren et.al. (2017) on the geometry of humans'
knowledge of navigation space.
}
\details{
In their research Warren et.al. (2017) conduct an experiment in which a total
of 20 participants had to navigate through one of two versions of a virtual
maze using virtual reality headsets. One is the standard version, Euclidean
maze, with several objects hidden inside and shortcuts to these objects. The
other maze is exactly the same apart from that it contains wormholes by which
participants can be 'teleported' from one place in the maze to another, the
non-Euclidean maze.

The task participants had to execute was to walk to the several target
objects in the maze from a certain start object. There were two types of
targets, probe and standard and for each of the types there were 4 different
ones resulting in a total of 8 targets.

Before the actual experiment participants first completed a training phase in
which participants were trained to walk to each target object.

The variable of interest in this experiment was the angular error, the angle
between the intial walking direction of the participant and the target's
location (for specifics see article by Warren et.al.).

The angular error is a circular variable and can be described and analyzed
using circular statistics.
}
\keyword{datasets}
