% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mmme}
\alias{mmme}
\title{Create model matrices for a circular mixed-effects regression model}
\usage{
mmme(pred.I, data, pred.II)
}
\arguments{
\item{pred.I}{model equation for effects of component 1}

\item{data}{the dataframe used for analysis}

\item{pred.II}{model equation for effects of component 2}
}
\description{
Create model matrices for a circular mixed-effects regression model
}
