% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mmr}
\alias{mmr}
\title{Create model matrices circular regression}
\usage{
mmr(pred.I, data, pred.II)
}
\arguments{
\item{pred.I}{model equation for effects of component 1}

\item{data}{the dataframe used for analysis}

\item{pred.II}{model equation for effects of component 2}
}
\description{
Create model matrices circular regression
}
