% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mode_est}
\alias{mode_est}
\title{Compute the mode of a vector of linear data}
\usage{
mode_est(x)
}
\arguments{
\item{x}{a vector of linear data}
}
\description{
Compute the mode of a vector of linear data
}
