% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mode_est_circ}
\alias{mode_est_circ}
\title{Compute the mode of a vector of circular data}
\usage{
mode_est_circ(x)
}
\arguments{
\item{x}{a vector of circular data}
}
\description{
Compute the mode of a vector of circular data
}
\examples{
library(bpnreg)
mode_est_circ(subset(Motor, Cond = "exp")$Phaserad)

}
