% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvrnorm_arma_eigen}
\alias{mvrnorm_arma_eigen}
\title{Sample from a multivariate normal distribution}
\usage{
mvrnorm_arma_eigen(n, mu, sigma)
}
\arguments{
\item{n}{An integer indicating the number of samples to take.}

\item{mu}{A mean vector.}

\item{sigma}{A variance-covariance matrix.}
}
\description{
Sample from a multivariate normal distribution
}
\keyword{internal}
