% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{Omega.f}
\alias{Omega.f}
\title{Sample precision matrix}
\usage{
Omega.f(b, B, v, q, N)
}
\arguments{
\item{b}{subject specific random effects vectors}

\item{B}{prior sum of squares matrix, scale parameter Wishart distribution
(of b=random effect if prior close to 0-->no random effect)}

\item{v}{prior df=dimension Z}

\item{q}{dimension Z(number of random effects)}

\item{N}{sample size at second level}
}
\description{
Sample precision matrix
}
