% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pnme}
\alias{pnme}
\title{A Gibbs sampler for a projected normal mixed-effects model}
\usage{
pnme(
  theta_cos,
  theta_sin,
  X1,
  X2,
  Z1,
  Z2,
  ZtZ1,
  ZtZ2,
  R,
  pred,
  its,
  lag,
  burn,
  N
)
}
\arguments{
\item{theta_cos}{A List with the cosine of the circular dependent variable.}

\item{theta_sin}{A List with the sine of the circular dependent variable.}

\item{X1}{A list of fixed effect model matrices for component I.}

\item{X2}{A list of fixed effect model matrices for component II.}

\item{Z1}{A list of random effect model matrices for component I.}

\item{Z2}{A list of random effect model matrices for component II.}

\item{ZtZ1}{A list of transformed random effect model matrices for component I.}

\item{ZtZ2}{A list of transformed random effect model matrices for component II.}

\item{R}{A list of starting values for R.}

\item{pred}{An empty list for likelihood computation.}

\item{its}{An integer specifying the number of iterations}

\item{lag}{An integer specifying the amount of lag.}

\item{burn}{An integer specifying the number of burn-in iterations.}

\item{N}{An integer specifying the number of burn-in iterations.}
}
\description{
A Gibbs sampler for a projected normal mixed-effects model
}
