% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pnr}
\alias{pnr}
\title{A Gibbs sampler for a projected normal regression model}
\usage{
pnr(theta, X1r, X2r, its, lag, burn)
}
\arguments{
\item{theta}{A vector with the circular dependent variable.}

\item{X1r}{A model matrix for component I.}

\item{X2r}{A model matrix for component II.}

\item{its}{An integer specifying the number of iterations}

\item{lag}{An integer specifying the amount of lag.}

\item{burn}{An integer specifying the number of burn-in iterations.}
}
\description{
A Gibbs sampler for a projected normal regression model
}
