% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classfunctions.R
\name{print.bpnme}
\alias{print.bpnme}
\title{Print output from a Bayesian circular mixed-effects model}
\usage{
\method{print}{bpnme}(x, ...)
}
\arguments{
\item{x}{a \code{bpnme object} obtained from the function
\code{bpnme()}}

\item{...}{further arguments passed to or from other methods}
}
\value{
a print of selected output from a Bayesian circular mixed-effects model
}
\description{
Prints selected output from a Bayesian circular mixed-effects model
}
\examples{
library(bpnreg)
fit.Maps <- bpnme(pred.I = Error.rad ~ Maze + Trial.type + Learn.c + (1|Subject),
data = Maps,
its = 100, burn = 10, n.lag = 3)
print(fit.Maps)

}
