% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classfunctions.R
\name{print.bpnr}
\alias{print.bpnr}
\title{Print output from a Bayesian circular regression model}
\usage{
\method{print}{bpnr}(x, ...)
}
\arguments{
\item{x}{a \code{bpnr object} obtained from the function
\code{bpnr()}}

\item{...}{further arguments passed to or from other methods}
}
\value{
a print of selected output from a Bayesian circular regression model
}
\description{
Prints selected output from a Bayesian circular regression model
}
