% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rho_circ}
\alias{rho_circ}
\title{Compute the mean resultant length of a vector of circular data}
\usage{
rho_circ(theta, units = "radians")
}
\arguments{
\item{theta}{a circular variable in radians or degrees.}

\item{units}{measurement units of the circular variable c("radians",
"degrees").}
}
\description{
Compute the mean resultant length of a vector of circular data
}
\examples{
library(bpnreg)
rho_circ(subset(Motor, Cond == "exp")$PhaseDiff, units = "degrees")

}
