% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{slice_r_me}
\alias{slice_r_me}
\title{Sample R (latent lengths)}
\usage{
slice_r_me(t, mu1, mu2, r)
}
\arguments{
\item{t}{current outcome value}

\item{mu1}{current predicted linear mean of the first component}

\item{mu2}{current predicted linear mean of the second component}

\item{r}{current value for r}
}
\description{
Sample R (latent lengths)
}
