% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{slice_rcpp}
\alias{slice_rcpp}
\title{A slice sampler for the latent lengths r}
\usage{
slice_rcpp(X1, X2, theta, b1, b2, n, r)
}
\arguments{
\item{X1}{A model matrix for component I.}

\item{X2}{A model matrix for component II.}

\item{theta}{A vector with the circular dependent variable.}

\item{b1}{A matrix containing the coefficients of componetn I for the current iteration.}

\item{b2}{A matrix containing the coefficients of componetn II for the current iteration.}

\item{n}{An integer indicating the sample size of the data.}

\item{r}{A matrix with the estimates of r of the previous iteration.}
}
\description{
A slice sampler for the latent lengths r
}
