% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summe}
\alias{summe}
\title{Compute summary and model fit statistics for the circular mixed-effects
regression model}
\usage{
summe(output, mm)
}
\arguments{
\item{output}{from a mixed-effects model}

\item{mm}{output from the function mmme()}
}
\description{
Compute summary and model fit statistics for the circular mixed-effects
regression model
}
\keyword{internal}
