% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sumr}
\alias{sumr}
\title{Compute summary and model fit statistics for the circular regression model}
\usage{
sumr(output, mm)
}
\arguments{
\item{output}{from the regression estimation function pnr()}

\item{mm}{output from the function mmr()}
}
\description{
Compute summary and model fit statistics for the circular regression model
}
\keyword{internal}
