% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classfunctions.R
\name{traceplot}
\alias{traceplot}
\title{Traceplots}
\usage{
traceplot(object, parameter = "SAM", variable)
}
\arguments{
\item{object}{an object used to select a method}

\item{parameter}{one of c("B1", "B2", "a.x", "a.c", "b.c", "SAM", "AS", "SSDO",
"circ.diff") to indicate for which parameter a traceplot is required}

\item{variable}{a character string with variable name(s) to indicate for
which variable(s) a traceplot is required}
}
\description{
Traceplot function for a \code{bpnr object} or \code{bpnme object}.
}
