% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{bpnme}
\alias{bpnme}
\title{Fit a Baysian circular mixed-effects model}
\source{
Cremers, J., Mainhard, M.T. & Klugkist, I. (2017). Assessing a
  Bayesian Embedding Approach to Circular Regression Models. Manuscript under
  review.

Cremers, J., & Klugkist, I. (2017). How to analyze circular data: A
  tutorial for projected normal regression models. Working paper.

Cremers, J., Mainhard, M.T. & Klugkist, I. (2017). Longitudinal
  circular modelling of circumplex measurements for teacher behavior. Working
  paper.

Presnell, B., Morrison, S.P. & Littell, R.C. (1998). Projected
  multivariate linear models for directional data. Journal of the Americal
  Statistical Association, 93 (443), 1068 - 1077.
}
\usage{
bpnme(pred.I, data, pred.II = pred.I, its = 1000, burn = 1, n.lag = 1,
  seed = NULL)
}
\arguments{
\item{pred.I}{model equation for effects of component 1.}

\item{data}{the dataframe used for analysis.}

\item{pred.II}{model equation for effects of component 2.}

\item{its}{output iterations of the MCMC sampler.}

\item{burn}{number of burn-in iterations.}

\item{n.lag}{amount of lag for the iterations and burn-in.}

\item{seed}{user-specified random seed.}
}
\value{
A \code{bpnme} object, which can be further analyzed using the
  associated functions \code{\link{traceplot.bpnme}}, \code{\link{BFc.bpnme}},
  \code{\link{coef_lin.bpnme}}, \code{\link{coef_circ.bpnme}}, \code{\link{coef_ran.bpnme}},
  \code{\link{residuals.bpnme}}, \code{\link{predict.bpnme}},
  \code{\link{fit.bpnme}} and \code{\link{print.bpnme}}.

  A \code{bpnr} object contains the following elements (some elements are not
  returned if not applicable)

  \describe{
  \item{\code{Beta.I}}{A vector of the fixed effects for the first component.}
  \item{\code{Beta.II}}{A vector of the fixed effects for the second component.}
  \item{\code{B.I}}{out.}
  \item{\code{B.II}}{out.}
  \item{\code{VCovI}}{The random effect variances in the first component.}
  \item{\code{VCovII}}{The random effect variances in the second component.}
  \item{\code{predictiva}}{out.}
  \item{\code{circular.ri}}{The circular random intercepts.}
  \item{\code{N}}{Number of observed cases.}
  \item{\code{its}}{Number of output iterations.}
  \item{\code{n.lag}}{Number of iterations to save.
  one in \code{lag} iterations will be saved. Set lag to 1 to save all iterations (default).}
  \item{\code{burn}}{Burn-in time for the MCMC sampler.}
  \item{\code{p1}}{Number of parameters in the first component.}
  \item{\code{p2}}{Number of parameters in the second component.}
  \item{\code{q1}}{out.}
  \item{\code{q2}}{out.}
  \item{\code{a.x}}{Describes the location of the
  inflection point of the regression curve on the axis of the circular outcome.}
  \item{\code{a.c}}{Describes the axis of the predictor.}
  \item{\code{b.c}}{Describes the slope of the tangent line at the inflection point.}
  \item{\code{SAM}}{A vector of circular regression slopes at the mean.}
  \item{\code{AS}}{A vector of average slopes of the circular regression.}
  \item{\code{SSDO}}{The signed shorted distance to the origin.}
  \item{\code{circ.diff}}{The circular difference found between categorical variables in each iteration.}
  \item{\code{CRSnum}}{Values for the circular random slope of continuoes variables.}
  \item{\code{CRScat}}{Values for the circular random slope of categorical variables.}
  \item{\code{CRS}}{Values for the circular random slope.}
  \item{\code{CRI}}{The mean resultant length of the random intercept, a measure of concentration.}
  \item{\code{Call}}{The matched call.}
  \item{\code{lin.coef.I}}{The mean, mode, standard deviation and confidence
  interval of the highest posterior density of the linear coefficients for \code{B1}.}
  \item{\code{lin.coef.II}}{The mean, mode, standard deviation and confidence
  interval of the highest posterior density of the linear coefficients for \code{B2}.}
  \item{\code{circ.coef}}{The mean, mode, standard deviation and confidence
  interval of the highest posterior density for the \code{a.x}, \code{a.c},
  \code{b.c}, \code{AS}, \code{SAM} and \code{SSDO} of the circular coefficients.}
  \item{\code{circ.coef.cat}}{The mean, mode, standard deviation and confidence
  interval of the highest posterior density of the categorical coefficients}
  \item{\code{circ.coef.means}}{The mean, mode, standard deviation and confidence
  interval of the highest posterior density of the categorical coefficients including an intercept.}
  \item{\code{model.fit}}{A list of information criteria for assesment of model fit.}
  \item{\code{lin.res.varrand.I}}{out.}
  \item{\code{lin.res.varrand.II}}{out.}
  \item{\code{circ.res.varrand}}{out.}
  \item{\code{mm}}{A list of information on the specified model.}
  }
}
\description{
This function fits a Bayesian circular mixed-effects model based on the
projected normal distribution.
}
\details{
Because the model is based on the projected normal distribution, a
  model equation has to be given for the fixed and random effects of thetwo
  components. By default the model equation of the second component
  \code{pred.II} is set to be equal to that of the
  first component. For more information about the projected normal
  distribution see Presnell, Morrisson & Littell (1998).

  A tutorial on how to use this function can be found in Cremers & Klugkist
  (2017, working paper). More details on the sampling algorithm and
  interpretation of the coefficients from the model can be found in Cremers,
  Mainhard & Klugkist (2017, working paper).
}
\examples{
library(bpnreg)
bpnme(Error.rad ~ Maze + Trial.type + (1|Subject), Maps, its = 100)

}
