% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{bpnr}
\alias{bpnr}
\title{Fit a Bayesian circular regression model}
\source{
Cremers, J., Mulder, K.T. & Klugkist, I. (2018). Circular
  interpretation of regression coefficients. British Journal of Mathematical
  and Statistical Psychology, 71(1), 75-95.

Cremers, J., Mainhard, M.T. & Klugkist, I. (in press). Assessing a
  Bayesian Embedding Approach to Circular Regression Models. Methodology.

Cremers, J., & Klugkist, I. (2017). How to analyze circular data: A
  tutorial for projected normal regression models. Under review.

Presnell, B., Morrison, S.P. & Littell, R.C. (1998). Projected
  multivariate linear models for directional data. Journal of the Americal
  Statistical Association, 93 (443), 1068 - 1077.
}
\usage{
bpnr(pred.I, data, pred.II = pred.I, its = 1000, burn = 1, n.lag = 1,
  seed = NULL)
}
\arguments{
\item{pred.I}{model equation for effects of component 1.}

\item{data}{the dataframe used for analysis.}

\item{pred.II}{model equation for effects of component 2.}

\item{its}{output iterations of the MCMC sampler.}

\item{burn}{number of burn-in iterations.}

\item{n.lag}{amount of lag for the iterations and burn-in.}

\item{seed}{user-specified random seed.}
}
\value{
A \code{bpnr} object, which can be further analyzed using the
  associated functions \code{\link{traceplot.bpnr}}, \code{\link{BFc.bpnr}},
  \code{\link{coef_lin.bpnr}}, \code{\link{coef_circ.bpnr}},
  \code{\link{residuals.bpnr}} ,\code{\link{predict.bpnr}},
  \code{\link{fit.bpnr}} and \code{\link{print.bpnr}}.

  A \code{bpnr} object contains the following elements (some elements are not
  returned if not applicable)

  \describe{
  \item{\code{B1}}{A matrix of posterior samples for the coefficients \code{B1}
  of the first component.}
  \item{\code{B2}}{A matrix of posterior samples for the coefficients \code{B2}
  for the second component.}
  \item{\code{Likelihood}}{A matrix containing the posterior density values for all individuals in the dataset
   for all iterations. The rowsums of this matrix are the likelihood values for all iterations}
  \item{\code{its}}{Number of output iterations.}
  \item{\code{n.lag}}{One in \code{n.lag} iterations will be saved as output iteration. Set lag to 1 to save all iterations (default).}
  \item{\code{burn-in}}{Burn-in time for the MCMC sampler.}
  \item{\code{p1}}{Number of parameters predicting the first component.}
  \item{\code{p2}}{Number of parameters predicting the second component.}
  \item{\code{theta}}{The circular outcome vector.}
  \item{\code{a.x}}{A matrix of posterior samples for \code{a.x} which describes the location of the
  inflection point of the regression curve on the axis of the predictor.}
  \item{\code{a.c}}{A matrix of posterior samples for \code{a.c} which describes the location of the
  inflection point of the regression curve on the axis of the circular outcome.}
  \item{\code{b.c}}{A matrix of posterior samples for \code{b.c} which describes the slope of the tangent line at the inflection point.}
  \item{\code{SAM}}{A matrix of posterior samples for the circular regression slopes at the mean.}
  \item{\code{AS}}{A matrix of posterior samples for the average slopes of the circular regression.}
  \item{\code{SSDO}}{A matrix of posterior samples for the signed shortest distance to the origin.}
  \item{\code{circ.diff}}{A matrix of posterior samples for the circular difference between levels of categorical variables and the intercept.}
  \item{\code{Call}}{The matched call.}
  \item{\code{lin.coef.I}}{The mean, mode, standard deviation and 95 % confidence
  interval of the highest posterior density of the linear coefficients for \code{B1}.}
  \item{\code{lin.coef.II}}{The mean, mode, standard deviation and 95 % confidence
  interval of the highest posterior density of the linear coefficients for \code{B2}.}
  \item{\code{circ.coef}}{The mean, mode, standard deviation and 95 % confidence
  interval of the highest posterior density for the \code{a.x}, \code{a.c},
  \code{b.c}, \code{AS}, \code{SAM} and \code{SSDO} of the circular coefficients.}
  \item{\code{circ.coef.cat}}{The mean, mode, standard deviation and 95 % confidence
  interval of the highest posterior density the circular difference between levels of categorical variables and the intercept.}
  \item{\code{circ.coef.means}}{The mean, mode, standard deviation and 95 % confidence
  interval of the highest posterior density of circular means of the categorical variables.}
  \item{\code{model.fit}}{A list of information criteria for assesment of model fit.}
  \item{\code{mm}}{A list of information, model matrices, sample size, etc. on the specified model.}
  }
}
\description{
This function fits a Bayesian circular regression model based on the
projected normal distribution.
}
\details{
Because the model is based on the projected normal distribution, a
  model equation has to be given for two components. By default the equation
  of the second component \code{pred.II} is set to be equal to that of the
  first component. For more information about the projected normal
  distribution see Presnell, Morrisson & Littell (1998).

  A tutorial on how to use this function can be found in Cremers & Klugkist
  (2017, working paper). More details on the sampling algorithm and
  interpretation of the coefficients from the model can be found in Cremers,
  Mulder & Klugkist (2018) and Cremers, Mainhard & Klugkist (in press).
}
\examples{
library(bpnreg)
bpnr(Phaserad ~ Cond + AvAmp, Motor)

}
