% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classfunctions.R
\name{coef_lin}
\alias{coef_lin}
\title{Linear coefficients}
\usage{
coef_lin(object)
}
\arguments{
\item{object}{an object used to select a method}
}
\description{
\code{coef_lin} gives posterior summaries of the linear coefficients
}
\details{
the methods \link[pnreg]{coef_lin.bpnr} and
  \link[pnreg]{coef_lin.bpnme} have their own help page
}
