% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DIC_reg}
\alias{DIC_reg}
\title{Compute Model Fit Measures Regression Model}
\usage{
DIC_reg(theta, beta1, beta2, Likelihood, X1, X2)
}
\arguments{
\item{theta}{circular outcome values}

\item{beta1}{regression coefficients for the second component for each mcmc iteration from pnr function}

\item{beta2}{regression coefficients for the second component for each mcmc iteration from pnr function}

\item{Likelihood}{likelihood values for each individual and mcmc itertion from pnr function}

\item{X1}{model matrix for the first component}

\item{X2}{model matrix for the second component}
}
\description{
Compute Model Fit Measures Regression Model
}
