% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classfunctions.R
\name{fit}
\alias{fit}
\title{Model fit}
\usage{
fit(object)
}
\arguments{
\item{object}{an object used to select a method}
}
\description{
\code{fit} gives several model fit statistics
}
\details{
the methods \link[pnreg]{fit.bpnr} and
  \link[pnreg]{fit.bpnme} have their own help page
}
