% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_circ}
\alias{mean_circ}
\title{Compute the mean of a vector of circular data}
\usage{
mean_circ(theta, units = "radians")
}
\arguments{
\item{theta}{a circular variable in radians.}

\item{units}{measurement units of the circular variable c("radians", "degrees").}
}
\description{
Compute the mean of a vector of circular data
}
\examples{
library(bpnreg)
mean_circ(subset(Motor, Cond == "exp")$PhaseDiff, units = "degrees")

}
