% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pnme}
\alias{pnme}
\title{A Gibbs sampler for a projected normal mixed-effects model}
\usage{
pnme(
  theta_cos,
  theta_sin,
  X1,
  X2,
  Z1,
  Z2,
  ZtZ1,
  ZtZ2,
  R,
  pred,
  its,
  lag,
  burn,
  N
)
}
\arguments{
\item{its}{An integer specifying the number of iterations}

\item{lag}{An integer specifying the amount of lag.}

\item{burn}{An integer specifying the number of burn-in iterations.}

\item{theta}{A List with the circular dependent variable.}

\item{X1r}{A list of random effect model matrices for component I.}

\item{X2r}{A list of random effect model matrices for component II.}
}
\description{
A Gibbs sampler for a projected normal mixed-effects model
}
