% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classfunctions.R
\name{print.bpnme}
\alias{print.bpnme}
\title{Print output from a Bayesian circular mixed-effects model}
\usage{
\method{print}{bpnme}(x, ...)
}
\arguments{
\item{x}{a \code{bpnme object} obtained from the function \code{bpnme()}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A print of selected output from a Bayesian circular mixed-effects
  model.
}
\description{
Prints selected output from a Bayesian circular mixed-effects model.
}
\examples{
library(bpnreg)
fit.Maps <- bpnme(pred.I = Error.rad ~ Maze + Trial.type + L.c + (1|Subject),
data = Maps,
its = 100, burn = 1, n.lag = 1)
print(fit.Maps)

}
