% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sd_circ}
\alias{sd_circ}
\title{Compute the standard deviation of a vector of circular data}
\usage{
sd_circ(theta, type = "radians")
}
\arguments{
\item{theta}{a circular variable in radians.}

\item{type}{measurement units of the circular variable c("radians", "degrees")}
}
\description{
Compute the standard deviation of a vector of circular data
}
