% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classfunctions.R
\name{traceplot.bpnme}
\alias{traceplot.bpnme}
\title{Traceplots for a Bayesian circular mixed-effects model}
\usage{
\method{traceplot}{bpnme}(object, parameter = "SAM", variable)
}
\arguments{
\item{object}{a \code{bpnme object} obtained from the function \code{bpnme()}}

\item{parameter}{one of c("Beta.I", "Beta.II", a.x", "a.c", "b.c", "SAM",
"AS", "SSDO", "circ.diff") to indicate for which parameter a traceplot is
required}

\item{variable}{if effect = "fixed", a character string with variable name(s)
to indicate for which variable(s) a traceplot is required}
}
\description{
General plot function for a \code{bpnme object}.
}
