% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classfunctions.R
\name{traceplot.bpnr}
\alias{traceplot.bpnr}
\title{Traceplots for a Bayesian circular regression model}
\usage{
\method{traceplot}{bpnr}(object, type = "trace", parameter = "SAM",
  variable)
}
\arguments{
\item{object}{a \code{bpnr object} obtained from the function \code{bpnr()}}

\item{parameter}{one of c("B1", "B2", "a.x", "a.c", "b.c", "SAM", "AS", "SSDO",
"circ.diff") to indicate for which parameter a traceplot is required}

\item{variable}{a character string with variable name(s) to indicate for
which variable(s) a traceplot is required}
}
\description{
General plot function for a \code{bpnr object}.
}
